requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Cors>=4.0.0
Flask-Login>=0.6.0
Flask-Migrate>=4.0.0
Flask-Marshmallow>=0.15.0
Flask-Limiter>=3.5.0
Flask-WTF>=1.2.0
marshmallow-sqlalchemy>=0.29.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
requests>=2.31.0
reportlab>=4.0.0
openpyxl>=3.1.0
email-validator>=2.0.0
